+++ sdk/storage/Azure.Storage.Files.DataLake/src/DataLakeClientOptions.cs
@@ -353,5 +353,11 @@ namespace Azure.Storage.Files.DataLake
         {
             return this.Build(credentials, GeoRedundantSecondaryUri);
         }
+
+        /// <summary>
+        /// Gets or sets the Audience to use for authentication with Azure Active Directory (AAD). The audience is not considered when using a shared key.
+        /// </summary>
+        /// <value>If <c>null</c>, <see cref="DataLakeAudience.PublicAudience" /> will be assumed.</value>
+        public DataLakeAudience? Audience { get; set; }
     }
 }