+++ sdk/storage/Azure.Storage.Files.Shares/src/ShareClientOptions.cs
@@ -265,5 +265,11 @@ namespace Azure.Storage.Files.Shares
             Diagnostics.LoggedQueryParameters.Add("copyid");
             Diagnostics.LoggedQueryParameters.Add("restype");
         }
+
+        /// <summary>
+        /// Gets or sets the Audience to use for authentication with Azure Active Directory (AAD). The audience is not considered when using a shared key.
+        /// </summary>
+        /// <value>If <c>null</c>, <see cref="ShareAudience.PublicAudience" /> will be assumed.</value>
+        public ShareAudience? Audience { get; set; }
     }
 }